deps: Allow text-2.0
text-2.0 has been released and everything seems to be working just fine.

I tested with it on GHC 8.10.7 and GHC 9.2.1. I could not test on 9.0.1
due to technical difficulties with haskell.nix, but I suppose it works.

The issue with haskell.nix is that text-2.0 broke an internal interface
used by hashable, so we need a new hashable, but it fails to build for
an unknown reason.