Trigger CI on push to master and on PR
Problem: CI is triggered only on push, so it won't work in case of
PRs from forks.
Solution: trigger it on PRs as well. Trigger on push only when the
branch is master.
If you want to run on push to your feature branch, you can edit
check.yml in your branch.